Cleaned upped some obsolete stuff.

+++ CellAO/AO.Servers/ZoneEngine/MessageHandlers/KnuBotAnswerHandler.cs
@@ -51,7 +51,7 @@ namespace ZoneEngine.MessageHandlers
             var npc =
                 (NonPlayerCharacterClass)
                 FindDynel.FindDynelById(
-                    (int)knuBotAnswerMessage.Target.Type, knuBotAnswerMessage.Target.Instance);
+                    knuBotAnswerMessage.Target.Type, knuBotAnswerMessage.Target.Instance);
             var character = FindClient.FindClientById(message.Header.Sender).Character;
             if (npc != null)
             {
